python/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymysql==1.1.1
pillow==11.3.0
numpy==2.2.6
requests==2.32.4
torch==2.8.0
transformers==4.55.4
python-multipart==0.0.20
